% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.r
\name{as.Model}
\alias{as.Model}
\title{Create a modelling 'package', that includes all parts that might be needed at modelling.
This means that all relevant modelling paramewters may be defined at one place.}
\usage{
as.Model(formula, start = NULL, postpr = NULL)
}
\description{
Create a modelling 'package', that includes all parts that might be needed at modelling.
This means that all relevant modelling paramewters may be defined at one place.
}
