% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-stuff.r
\name{check_available_columns}
\alias{check_available_columns}
\title{Report missing columns in a data.frame.}
\usage{
check_available_columns(df, required_cols = NULL, message = "")
}
\arguments{
\item{df}{The data.frame.}

\item{required_cols}{A character vector of the required column names.}

\item{message}{A prefix for error messages.}
}
\value{
A character vector of the missing column names.
}
\description{
Report missing columns in a data.frame.
}
\seealso{
read_csv_file, write_csv_file
}
\author{
Peder Axensten
}
