% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.r
\name{file.name}
\alias{file.name}
\title{Returns the file name without any suffix.}
\usage{
file.name(path)
}
\arguments{
\item{path, }{a string. The path (or similar) to the data.}
}
\value{
The file name in \code{path} without any suffix.
}
\description{
Returns the file name without any suffix.
}
\seealso{
file.suffix
}
\author{
Peder Axensten
}
