% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.r
\name{file.suffix}
\alias{file.suffix}
\title{Returns the file name suffix, including the leading ".".}
\usage{
file.suffix(path)
}
\arguments{
\item{path, }{a string. The path (or similar) to the data.}
}
\value{
The file name suffix of \code{path}, including the leading ".".
}
\description{
Returns the file name suffix, including the leading ".".
}
\seealso{
file.name
}
\author{
Peder Axensten
}
