% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.r
\name{fix_predicted_root}
\alias{fix_predicted_root}
\title{Bias correction for predicted squareroot values.
The correction factor is mean( true variable values )/mean( predicted variable values ).}
\usage{
fix_predicted_root(model, values, tag, plots = NULL)
}
\description{
Bias correction for predicted squareroot values.
The correction factor is mean( true variable values )/mean( predicted variable values ).
}
