% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata-stuff.r
\name{geodata.stuff}
\alias{geodata.stuff}
\alias{epsg_string}
\alias{read_one_raster}
\alias{read_raster_extent}
\alias{read_rasters}
\alias{write_rasters}
\title{A family of functions that implements simple logging.}
\usage{
epsg_string(epsg)

read_one_raster(source, extent = NULL, message = "")

read_raster_extent(source, message = "")

read_rasters(source, auxiliary_dir, names, extent = NULL, message = "")

write_rasters(destination_stub, rasters, variant)
}
\arguments{
\item{epsg}{The epsg id of the crs.}

\item{source}{The file path of the raster.}

\item{extent}{The part (bounding box, object of class SpatExtent) of the raster to read.}

\item{message}{A prefix for error messages.}

\item{destination_stub}{A path stub that the name of each raster in the list is added to.}

\item{rasters}{A list of named rasters.}

\item{dirs}{One or more directory paths, where to look for the file.}

\item{file}{The file name.}

\item{suffices}{One or more optional file name suffices, an alias for names[ i ] will be paste0( names[ i ], suffix ).}
}
\description{
A family of functions that implements simple logging.
}
\section{Functions}{
\itemize{
\item \code{epsg_string}: Construct a crs string from an epsg number. Returns a WKT2 string of the CRS for recent versions of PROJ, rgdal & co., 
otherwise a "+init=epsg:x" proj-string (proj-strings are being deprecated in PROJ, replaced by WKT2).

\item \code{read_one_raster}: Read one raster file. Returns a [terra] raster.

\item \code{read_raster_extent}: Returns the extension (bounding box) of a raster. Object of class SpatExtent.

\item \code{read_rasters}: Read one or more raster files into a list. Returns a list of metric rasters.

\item \code{write_rasters}: Write a list of named rasters.
}}

\references{
Kristian Evers and Thomas Knudsen, \emph{Transformation pipelines for PROJ.4,} 
    				\url{https://www.fig.net/resources/proceedings/fig_proceedings/fig2017/papers/iss6b/ISS6B_evers_knudsen_9156.pdf} (2017)

\emph{SpatialLite and PROJ.6,} 
    				\url{https://www.gaia-gis.it/fossil/libspatialite/wiki?name=PROJ.6}

Roger Bivand, \emph{ECS530: (III) Coordinate reference systems,} 
   				\url{https://rsbivand.github.io/ECS530_h19/ECS530_III.html} (2019-12-03)

Edzer Pebesma and Roger Bivand, \emph{R spatial follows GDAL and PROJ development,} 
				\url{https://www.r-spatial.org/r/2020/03/17/wkt.html#crs-objects-in-sf} (2020-03-17)
}
\seealso{
\code{\link{pax.regression.package}}
}
