% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.r
\name{get_ruta_id}
\alias{get_ruta_id}
\title{Get the ruta id.}
\usage{
get_ruta_id(metrics_path_stub)
}
\arguments{
\item{metrics_path_stub, }{a path stub to the metrics files. Also defines the ruta id.}
}
\value{
The corresponding ruta's id.
}
\description{
Get the ruta id.
}
