% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{logging}
\alias{logging}
\alias{log_set_sink}
\alias{log_item}
\alias{log_stop}
\alias{log_sink}
\alias{log_begin}
\alias{log_end}
\title{A family of functions that implements simple logging.}
\usage{
log_set_sink(sink = nullfile())

log_item(...)

log_stop(..., call. = FALSE, domain = NULL)

log_sink()

log_begin()

log_end()
}
\arguments{
\item{sink}{Any textual sink, such as a file name, \code{nullfile()}, \code{stdout()}, \code{stderr()}, etc. 
It is saved and used in following calls to log functions. A value of \code{nullfile()} (the default value) will turn logging off.}

\item{...}{Error message. It will be logged as \code{paste0(...)}.}

\item{call.}{Past on to stop: logical, indicating if the call should become part of the error message.}

\item{domain}{Past on to stop: see \code{gettext}. If \code{NA}, messages will not be translated.}
}
\description{
A family of functions that implements simple logging.
}
\section{Functions}{
\itemize{
\item \code{log_set_sink}: Set (or reset) the logging sink (destination).

\item \code{log_item}: Log an item.

\item \code{log_stop}: An error occured, log the error message and stop.

\item \code{log_sink}: Log Returns the logging sink (destination).

\item \code{log_begin}: Log the entry into a function.

\item \code{log_end}: Log the exit from a function.
}}

\seealso{
\code{\link{pax.regression.package}}
}
