% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-base.r
\name{model_and_predict}
\alias{model_and_predict}
\title{Implements basic regression.}
\usage{
model_and_predict(
  variant,
  source_stub,
  plots,
  dest_dir,
  auxiliary_data_dir,
  number_of_candidates = 350,
  metadata_dest_dir = NA
)
}
\arguments{
\item{variant}{(string) What kind of modelling to do. At the moment, only "basic_linear" is implemented.}

\item{source_stub}{A path stub to the metrics files.}

\item{plots}{A path or a data.fram of the data to process.}

\item{dest_dir}{A path + file prefix, used for all file output.}

\item{auxiliary_data_dir}{A path to the auxiliary data directory.}

\item{number_of_candidates}{(unsigned integer) The number of plot kandidates to supply to the outlier removal.}

\item{metadata_output}{Produces process metadata files if true.}
}
\description{
\subsection{Want to implement more models or variants?}{
In this package, S3 dispatch is used. All you need to implement is the functions mention below, 
no changes in the base code such as this function should be necessary. 
If you want to implement other types of models, do it by following the 'basic_linear' case.
All relevant functions are implemented as S3 classes, so implement them as 
\enumerate{
- \code{independent_variables.my_model()},
- \code{dependent_preparation.my_model()}, 
- \code{build_models.my_model()}, and
- \code{predict.my_model()}.
}
After that you can call this function with the argument \code{variant="my_model"}.
If you want to implement an entirely different regression method, the you can reimplement this
function as \code{model_and_predict.my_model()}. But this default implementation should suffice for most cases.
}
}
\author{
Peder Axensten
}
