% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-base.r
\name{modelling_details}
\alias{modelling_details}
\alias{dependent_variables}
\alias{dependent_preparation}
\alias{independent_variables}
\alias{independent_preparation}
\alias{build_models}
\title{A bunch of functions that implements simple logging.}
\usage{
dependent_variables(variant)

dependent_preparation(variant, x)

independent_variables(variant)

independent_preparation(variant, x)

build_models(variant, plots, auxiliary_data_dir, metadata_dest_dir = NA)
}
\arguments{
\item{variant}{The object used for S3 dispatch.}

\item{x}{A data.frame of the data to process.}

\item{plots}{A data.fram of the data to process.}

\item{auxiliary_data_dir}{Where to find auxiliary data, if needed.}

\item{metadata_dest_dir}{Store metadata here, if needed.}
}
\description{
A bunch of functions that implements simple logging.
}
\section{Functions}{
\itemize{
\item \code{dependent_variables}: Returns a vector of strings, with the names of the dependent variables to calculate.

\item \code{dependent_preparation}: Returns the data.frame \code{x}, with whatever additional columns.

\item \code{independent_variables}: Returns a vector of strings, with the names of the the independent variables.

\item \code{independent_preparation}: Calculate data needed by the modelling AND extra rasters before using the models.
Returns the data.frame \code{x}, with whatever additional columns.

\item \code{build_models}: Calculate the models.
}}

\seealso{
\code{\link{pax.regression.package}}
}
