% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pax-regression-package.r
\docType{package}
\name{pax.regression.package}
\alias{pax.regression.package}
\title{Tools to use for regression in the forestry domain.}
\description{
This package implements various regression methods.
}
\details{
\subsection{Regression methods}{
\enumerate{
\item \code{\link{basic_linear}} basic linear regression using rlm.
\item \code{\link{lidar_exponential}} non-linear regression using nls.
}}
}
\seealso{
\code{\link{pax.regression}} To choose plots, remove outliers, build models, and predict.

\code{\link{plot_sample_ruta}} To choose plots.

\code{\link{plot_sample_outlier_reduction}} Remove outliers among ploits.

\code{\link{variant_alts}} List what estimation methods that are available.
}
\author{
Peder Axensten
}
