% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pax-regression-package.r
\name{pax.regression}
\alias{pax.regression}
\title{Calls the 'variant' regression function.}
\usage{
pax.regression(
  source,
  plots,
  dest,
  auxiliary_data_dir,
  number_of_candidates = 350,
  outlier_removal_perc = 10,
  metadata_dest_dir_dir = NULL
)
}
\arguments{
\item{source}{A path stub to the metrics files. Also defines the ruta id.}

\item{plots}{A path to a .csv file or a data.frame of the plots.}

\item{dest}{A path + file prefix, used for all file output.}

\item{auxiliary_data_dir}{A path to the auxiliary data directory.}

\item{number_of_candidates}{(unsigned integer) The number of plot kandidates to supply to the outlier removal.}

\item{outlier_removal_perc}{(float in [0,95]) The percentage of plot candidates to remove in thge outlier removal.}

\item{metadata_dest_dir}{If no is.na metadata is stored here.}
}
\description{
Calls the 'variant' regression function.
}
\seealso{
\code{\link{pax.regression.package}}
}
\author{
Peder Axensten
}
