% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sample-outlier-reduction.r
\name{plot_sample_outlier_reduction}
\alias{plot_sample_outlier_reduction}
\title{Implements outlier reductiuon on a forest plots table.}
\usage{
plot_sample_outlier_reduction(
  plots,
  outlier_removal_perc,
  metadata_dest_dir = NA
)
}
\arguments{
\item{plots}{A path to a .csv file or a data.frame of the plots.}

\item{outlier_removal_perc}{(float in [0,95]) The percentage of plot candidates to remove in thge outlier removal.}

\item{metadata_dest_dir}{If no is.na metadata is stored here.}
}
\value{
A subset of plots.
}
\description{
Implements outlier reductiuon on a forest plots table.
}
\seealso{
\code{\link{pax.regression.package}}
}
\author{
Peder Axensten
}
