% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sample-ruta.r
\name{plot_sample_ruta}
\alias{plot_sample_ruta}
\title{Given a table of all plots, it finds an optimal sub-sample of plots for the given file id.}
\usage{
plot_sample_ruta(
  ruta_id,
  plots,
  ruta_metadata,
  number_of_candidates,
  metadata_dest_dir = NA
)
}
\arguments{
\item{ruta_id, }{.}

\item{plots}{A path to a .csv file or a data.frame of the plots.}

\item{number_of_candidates}{(unsigned integer) The number of plot kandidates to supply to the outlier removal.}

\item{metadata_dest_dir}{If no is.na metadata is stored here.}

\item{auxiliary_data_dir}{A path to the auxiliary data directory.}

\item{k_min}{Exclude plots with hgv/p95 < k_min.}

\item{k_max}{Exclude plots with hgv/p95 > k_max.}
}
\value{
A data.frame containing a subset of the lines of the original plots table.
}
\description{
Given a table of all plots, it finds an optimal sub-sample of plots for the given file id.
}
\seealso{
\code{\link{pax.regression.package}}
}
\author{
Peder Axensten
}
