% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-stuff.r
\name{read_csv_file}
\alias{read_csv_file}
\title{Read a csv-like file and return a data.frame. If the argument is a data.frame, just return it.}
\usage{
read_csv_file(path_or_df, required_cols = NULL, message = "")
}
\arguments{
\item{path_or_df}{A string. The path or data.frame of the data.}

\item{required_cols}{A character vector of required column names.}

\item{message}{A prefix for error messages.}
}
\value{
The data.frame.
}
\description{
The function does some checks:
- If \code{path_or_df} is a character string it is treated as a path. If there is no such file
  an error is reported.
- If \code{required_cols} is defined, any missing column names are reported as an error.
}
\seealso{
write_csv_file, check_available_columns
}
\author{
Peder Axensten
}
