% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sample-outlier-reduction.r
\name{required_plot_cols.outlier_reduction}
\alias{required_plot_cols.outlier_reduction}
\title{Get a vector of strings, with the names of the required variables used in the outlier reduction.}
\usage{
required_plot_cols.outlier_reduction()
}
\description{
Get a vector of strings, with the names of the required variables used in the outlier reduction.
}
\author{
Peder Axensten
}
