% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sample-ruta.r
\name{required_plot_cols.plot_sample_ruta}
\alias{required_plot_cols.plot_sample_ruta}
\title{Get a vector of strings, with the names of the required variables used in the candidate selection.}
\usage{
required_plot_cols.plot_sample_ruta()
}
\value{
A vector of strings, with the names of the required variables used in the candidate selection.
}
\description{
Get a vector of strings, with the names of the required variables used in the candidate selection.
}
\author{
Peder Axensten
}
