% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.r
\name{variant_alts}
\alias{variant_alts}
\title{Report a vector of modelling variants.}
\usage{
variant_alts()
}
\description{
Report a vector of modelling variants.
}
\seealso{
\code{\link{pax.regression.package}}
}
