% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-stuff.r
\name{write_csv_file}
\alias{write_csv_file}
\title{Write a data.frame or similar object as a csv-like file.}
\usage{
write_csv_file(df, dest_path)
}
\arguments{
\item{df}{A data.frame or similar object.}

\item{dest_path}{A string. The destina<tion path.}
}
\value{
The data.frame.
}
\description{
Write a data.frame or similar object as a csv-like file.
}
\seealso{
read_csv_file, check_available_columns
}
\author{
Peder Axensten
}
